✨ email subscriptions (#453)
* adds environment variable for mailchimp subscription list

* uncomments existing work

* includes opt-in flags in user model

* adds REST call to persona for subscription

fixes the history not being passed in

* hide nih subscription for now

* enables the nih email subscription

* disable button if has not consented

* get rid of buffer prop
