+++ .env.schema
@@ -48,3 +48,4 @@ REACT_APP_KF_WEB_ROOT=
 REACT_APP_KF_FACEBOOK=
 REACT_APP_KF_TWITTER=
 REACT_APP_KF_GITHUB=
+REACT_APP_KF_MAILCHIMP_LIST=
