+++ src/common/injectGlobals.js
@@ -59,3 +59,4 @@ export const kfWebRoot: string = process.env.REACT_APP_KF_WEB_ROOT;
 export const kfFacebook: string = process.env.REACT_APP_KF_FACEBOOK;
 export const kfTwitter: string = process.env.REACT_APP_KF_TWITTER;
 export const kfGithub: string = process.env.REACT_APP_KF_GITHUB;
+export const kfMailchimpList: string = process.env.REACT_APP_KF_MAILCHIMP_LIST;
