+++ src/uikit/Row.js
@@ -4,7 +4,6 @@ import { justifyContent, space } from 'styled-system';
 
 export default styled(Flex)`
   ${({ theme }) => theme.row};
-  ${({ buffer }) => (buffer ? `margin-top: 10px;` : ``)};
   ${({ center, theme }) => (center ? theme.center : ``)};
   ${justifyContent};
   ${space};
